fix params and docblocks

+++ classes/phing/listener/statistics/StatisticsReport.php
@@ -81,7 +81,7 @@ class StatisticsReport
         $total = count($totalTimes);
         for ($i = 0; $i < $total; $i++) {
             $totalTime = $totalTimes[$i];
-            $round     = round(100 * (double) $totalTime / $runningTotalTime);
+            $round     = round(100 * (float) $totalTime / $runningTotalTime);
             $table->put($i + 1, self::$IDX_PERCENTAGE, (string) $round);
         }
     }
